requirements.txt
pytube
openai-whisper
assemblyai
bardapi
streamlit-chat
fastapi
uvicorn
streamlit
requests
psycopg2-binary
openai
langchain
tiktoken
langchain==0.0.177
python-dotenv==1.0.0
sentence-transformers==2.2.2
pinecone-client==2.2.1
python-multipart
flask_cors